+++ app/build.gradle.kts
@@ -59,9 +59,9 @@ dependencies {
     implementation(libs.flexbox)
     implementation(libs.circle.indicator)
     implementation(libs.physicslayout)
-    implementation(libs.room.runtime)
+
     ksp(libs.room.compiler)
-    implementation(libs.room.ktx)
+    implementation(libs.bundles.room)
     implementation(libs.encrypted.sharedpreferences)
 
     implementation(libs.bundles.ktor)
