+++ app/src/main/java/com/davay/android/core/data/database/entity/SessionEntity.kt
@@ -1,4 +1,4 @@
-package com.davay.android.data.database.entity
+package com.davay.android.core.data.database.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
