+++ app/src/main/java/com/davay/android/core/data/impl/UserDataRepositoryImpl.kt
@@ -1,7 +1,7 @@
-package com.davay.android.data.impl
+package com.davay.android.core.data.impl
 
 import android.content.SharedPreferences
-import com.davay.android.domain.repositories.UserDataRepository
+import com.davay.android.core.domain.api.UserDataRepository
 import java.util.UUID.randomUUID
 
 class UserDataRepositoryImpl(
