--- app/src/main/java/com/davay/android/core/di/storage/model/PreferencesStorage.kt
@@ -1,5 +1,0 @@
-package com.davay.android.core.di.storage.model
-
-enum class PreferencesStorage {
-    USER
-}
