+++ app/src/main/java/com/davay/android/core/domain/usecases/SetUserDataUseCase.kt
@@ -1,0 +1,7 @@
+package com.davay.android.core.domain.usecases
+
+import com.davay.android.core.domain.models.UserDataFields
+
+interface SetUserDataUseCase {
+    fun setUserData(value: UserDataFields)
+}
