+++ app/src/main/java/com/davay/android/di/prefs/model/PreferencesStorage.kt
@@ -1,0 +1,5 @@
+package com.davay.android.di.prefs.model
+
+enum class PreferencesStorage {
+    USER
+}
