--- app/src/main/java/com/davay/android/domain/usecases/GetUserDataUseCase.kt
@@ -1,7 +1,0 @@
-package com.davay.android.domain.usecases
-
-import com.davay.android.domain.models.UserDataFields
-
-interface GetUserDataUseCase {
-    fun getUserData(userData: UserDataFields): String
-}
