--- app/src/main/java/com/davay/android/domain/usecases/SetUserDataUseCase.kt
@@ -1,7 +1,0 @@
-package com.davay.android.domain.usecases
-
-import com.davay.android.domain.models.UserDataFields
-
-interface SetUserDataUseCase {
-    fun setUserData(value: UserDataFields)
-}
