+++ app/src/main/java/com/davay/android/feature/main/di/MainDataModule.kt
@@ -1,12 +1,12 @@
 package com.davay.android.feature.main.di
 
 import android.content.SharedPreferences
-import com.davay.android.core.di.storage.marker.StorageMarker
-import com.davay.android.core.di.storage.model.PreferencesStorage
-import com.davay.android.data.impl.UserDataRepositoryImpl
-import com.davay.android.domain.impl.GetUserDataUseCaseImpl
-import com.davay.android.domain.repositories.UserDataRepository
-import com.davay.android.domain.usecases.GetUserDataUseCase
+import com.davay.android.core.data.impl.UserDataRepositoryImpl
+import com.davay.android.core.domain.api.UserDataRepository
+import com.davay.android.core.domain.impl.GetUserDataUseCaseImpl
+import com.davay.android.core.domain.usecases.GetUserDataUseCase
+import com.davay.android.di.prefs.marker.StorageMarker
+import com.davay.android.di.prefs.model.PreferencesStorage
 import dagger.Module
 import dagger.Provides
