+++ app/src/main/java/com/davay/android/feature/registration/presentation/RegistrationViewModel.kt
@@ -2,8 +2,8 @@ package com.davay.android.feature.registration.presentation
 
 import android.text.Editable
 import com.davay.android.base.BaseViewModel
-import com.davay.android.domain.models.UserDataFields
-import com.davay.android.domain.usecases.SetUserDataUseCase
+import com.davay.android.core.domain.models.UserDataFields
+import com.davay.android.core.domain.usecases.SetUserDataUseCase
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import javax.inject.Inject
